Implemented even more instructions

+++ LibChip32/IInstruction.cs
@@ -10,6 +10,6 @@ namespace LibChip8
     {
         public byte InstructionClassByte { get; }
         public byte InstructionIdentByte { get; }
-        public void Execute(CPU cpu, ulong instr);
+        public void Execute(CPU cpu, uint instr);
     }
 }
